refactor: Re-export types
